data/pipelines/iot_streaming/pipeline_iot_streaming_kafka_local/requirements_massive_simulation.txt
# Requirements pour Simulateur Massif IoT Patients KIDJAMO
# Installation: pip install -r requirements_massive_simulation.txt

# Core dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0

# IoT & Messaging
kafka-python>=2.0.2
twilio>=8.5.0
requests>=2.31.0

# Data processing
asyncio-mqtt>=0.13.0
pydantic>=2.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
pyyaml>=6.0

# Optional: Advanced analytics
scipy>=1.11.0
scikit-learn>=1.3.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
